requirements.txt
streamlit
streamlit-lottie
requests
pillow
streamlit-drawable-canvas
